Refactor user schema from types
