+++ backend/src/user/user.types.ts
@@ -1,35 +1,6 @@
 import { users } from "../db/db.schema.ts";
+import { createUserSchema, PublicUserSchema, userIdSchema, userNameSchema } from "./user.schema.ts";
 import { z } from "zod";
-
-export const createUserSchema = z
-    .object({
-        username: z.string().min(3, { message: "Username must be at least 3 characters long" }),
-        displayName: z
-            .string()
-            .min(3, { message: "Display name must be at least 3 characters long" }),
-        password: z.string().min(8, { message: "Password must be at least 8 characters long" }),
-        confirmPassword: z
-            .string()
-            .min(8, { message: "Confirm password must be at least 8 characters long" }),
-    })
-    .refine((data) => data.password === data.confirmPassword, {
-        path: ["confirmPassword"],
-        message: "Passwords do not match",
-    });
-
-export const userIdSchema = z.object({
-    id: z.coerce.number().int().gt(0),
-});
-
-export const userNameSchema = z.object({
-    username: z.string().min(3, { message: "Username is required" }),
-});
-
-// export const updateUserSchema = z.object({
-//     id: z.coerce.number().int().gt(0),
-//     username: z.string().min(3, { message: "Username must be at least 3 characters long" }),
-//     displayName: z.string().min(3, { message: "Display name must be at least 3 characters long" }),
-// });
 
 export type CreateUserDTO = z.infer<typeof createUserSchema>;
 export type UserIdDTO = z.infer<typeof userIdSchema>;
@@ -37,3 +8,4 @@ export type UserNameDTO = z.infer<typeof userNameSchema>;
 
 export type User = typeof users.$inferSelect;
 export type NewUser = typeof users.$inferInsert;
+export type PublicUser = z.infer<typeof PublicUserSchema>;
